refactor: pgdbConnection
