+++ .vscode/settings.json
@@ -1,3 +1,9 @@
 {
-  "cSpell.words": ["fortunetellerid", "mondooduang", "packageid", "tailwindcss"]
+  "cSpell.words": [
+    "fortunetellerid",
+    "mondooduang",
+    "packageid",
+    "pgdb",
+    "tailwindcss"
+  ]
 }
